+++ src/frontend/modules/EditorPanels/LayerSelectionPanel.css
@@ -1,0 +1,44 @@
+.layer-selection-container {
+  overflow: auto;
+  min-height: 100px;
+  display: flex;
+  flex-direction: column;
+}
+
+.layer-selection-panel {
+  border: 1px solid #ccc;
+  padding: 5px;
+  margin: 5px;
+}
+
+.layer-selection-header {
+  font-size: 12px;
+  padding: 3px;
+  border-bottom: 1px solid #ddd;
+  margin-bottom: 3px;
+}
+
+.layer-selection-list {
+  background-color: #f5f5f5;
+  padding: 8px;
+  border: 1px solid #ddd;
+  height: 450px;
+  overflow-y: scroll;
+  font-size: 12px;
+}
+
+.shape {
+  padding: 8px;
+  margin-bottom: 5px;
+  border: 1px solid #468dee;
+  border-radius: 0px;
+  background-color: #97c2fc;
+  cursor: grab;
+  width: 100%;
+}
+
+.parameter-list {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
