+++ src/frontend/modules/EditorPanels/NetworkCanvas.jsx
@@ -1,0 +1,14 @@
+const NetworkCanvas = ({ networkRef, handleDrop }) => {
+  return (
+    <div className="network-canvas">
+      <div
+        id="mynetwork"
+        ref={networkRef}
+        onDragOver={(e) => e.preventDefault()}
+        onDrop={handleDrop}
+      ></div>
+    </div>
+  );
+};
+
+export default NetworkCanvas;
